+++ src/lib/rt/src/rt_rtt_cmdmode.c
@@ -209,7 +209,7 @@ int	rtt_replace_symbol( char *command, char *newcommand)
 	char	*t;
 	char	*u;
 	int	symbolmode;
-	int	size;
+	long int size;
 	char	value[80];
 	char	symbol[80];
 	char	upper_symbol[80];
@@ -238,7 +238,7 @@ int	rtt_replace_symbol( char *command, char *newcommand)
 	    if ( symbolmode)
 	    {
 	      /* End of potential symbol */
-	      size = (int) s - (int) u;
+	      size = (long int) s - (long int) u;
 	      strncpy( symbol, u, size);
 	      symbol[size] = 0;
 	      rtt_toupper( upper_symbol, symbol);
