+++ wb/lib/wb/src/wb_goenm15.cpp
@@ -86,7 +86,7 @@ int goen_create_nodetype_m15(
     pwr_tClassId	cid,
     ldh_tSesContext	ldhses,
     flow_tCtx		ctx,
-    unsigned long 	*mask,
+    unsigned int 	*mask,
     unsigned long	subwindowmark,
     unsigned long	node_width,
     flow_tNodeClass	*node_class,
@@ -95,9 +95,9 @@ int goen_create_nodetype_m15(
   int  		i;
   int		inputpoints, outputpoints;
   unsigned long   pointmask;
-  unsigned long   *inmask_pointer;
-  unsigned long   *outmask_pointer;
-  unsigned long   *invertmask_pointer;
+  unsigned int   *inmask_pointer;
+  unsigned int   *outmask_pointer;
+  unsigned int   *invertmask_pointer;
   int 		inputs;
   int 		interns;
   int		outputs;
@@ -380,14 +380,14 @@ int goen_create_nodetype_m15(
 *
 **************************************************************************/
 int goen_get_point_info_m15( WGre *grectx, pwr_sGraphPlcNode *graphbody, 
-			    unsigned long point, unsigned long *mask, 
+			    unsigned long point, unsigned int *mask, 
 			    unsigned long node_width, goen_conpoint_type *info_pointer, 
 			    vldh_t_node node)
 {
     	int	inputpoints, outputpoints;
     	unsigned long    pointmask;
-	unsigned long    *inmask_pointer;
-	unsigned long    *outmask_pointer;
+	unsigned int    *inmask_pointer;
+	unsigned int    *outmask_pointer;
         double		ll_x,ll_y,ur_x,ur_y;
 	int		first_input, second_input;
 
@@ -462,17 +462,17 @@ int goen_get_point_info_m15( WGre *grectx, pwr_sGraphPlcNode *graphbody,
 **************************************************************************/
 int	goen_get_parameter_m15( pwr_sGraphPlcNode *graphbody, pwr_tClassId cid, 
 			       ldh_tSesContext ldhses, unsigned long con_point, 
-			       unsigned long *mask, unsigned long *par_type, 
+			       unsigned int *mask, unsigned long *par_type, 
 			       unsigned long *par_inverted, unsigned long *par_index)
 {
 
 	unsigned long	inputs,interns,outputs;
 	unsigned long    pointmask;
-	unsigned long    *inmask_pointer;
-	unsigned long    *outmask_pointer;
+	unsigned int     *inmask_pointer;
+	unsigned int     *outmask_pointer;
 	int		input_found, output_found;
 	int		inputpoints, outputpoints;
-	unsigned long	*invertmask_pointer;
+	unsigned int	*invertmask_pointer;
 	int		first_input, second_input;
 
 	/* Get number of parameters */
@@ -551,7 +551,7 @@ int	goen_get_parameter_m15( pwr_sGraphPlcNode *graphbody, pwr_tClassId cid,
 *
 **************************************************************************/
 int goen_get_location_point_m15( WGre *grectx, pwr_sGraphPlcNode *graphbody, 
-				unsigned long *mask, unsigned long node_width, 
+				unsigned int *mask, unsigned long node_width, 
 				goen_point_type *info_pointer, vldh_t_node node)
 {
